Reorganize special_cases.rs, so that codegen_special_cases.rs is private
Also modifies an invocation to use is_class_deleted() instead of is_class_excluded() -- changes semantics.
