+++ godot-codegen/src/special_cases/codegen_special_cases.rs
