+++ godot-codegen/src/special_cases/special_cases.rs
@@ -23,13 +23,15 @@
 
 #![allow(clippy::match_like_matches_macro)] // if there is only one rule
 
-use crate::codegen_special_cases;
 use crate::models::domain::TyName;
 use crate::models::json::{JsonBuiltinMethod, JsonClassMethod, JsonUtilityFunction};
+use crate::special_cases::codegen_special_cases;
 use crate::Context;
 
+// Deliberately private -- all checks must go through `special_cases`.
+
 #[rustfmt::skip]
-pub(crate) fn is_class_method_deleted(class_name: &TyName, method: &JsonClassMethod, ctx: &mut Context) -> bool {
+pub fn is_class_method_deleted(class_name: &TyName, method: &JsonClassMethod, ctx: &mut Context) -> bool {
     if codegen_special_cases::is_class_method_excluded(method, ctx){
         return true;
     }
@@ -49,12 +51,12 @@ pub(crate) fn is_class_method_deleted(class_name: &TyName, method: &JsonClassMet
     }
 }
 
-pub(crate) fn is_class_deleted(class_name: &TyName) -> bool {
+pub fn is_class_deleted(class_name: &TyName) -> bool {
     codegen_special_cases::is_class_excluded(&class_name.godot_ty)
         || is_godot_type_deleted(class_name)
 }
 
-pub(crate) fn is_godot_type_deleted(ty_name: &TyName) -> bool {
+pub fn is_godot_type_deleted(ty_name: &TyName) -> bool {
     // Exclude experimental APIs unless opted-in.
     if !cfg!(feature = "experimental-godot-api") && is_class_experimental(ty_name) {
         return true;
@@ -154,7 +156,7 @@ fn is_class_experimental(class_name: &TyName) -> bool {
 
 /// Whether a method is available in the method table as a named accessor.
 #[rustfmt::skip]
-pub(crate) fn is_named_accessor_in_table(class_or_builtin_ty: &TyName, godot_method_name: &str) -> bool {
+pub fn is_named_accessor_in_table(class_or_builtin_ty: &TyName, godot_method_name: &str) -> bool {
     // Hand-selected APIs.
     match (class_or_builtin_ty.godot_ty.as_str(), godot_method_name) {
         | ("OS", "has_feature")
@@ -169,7 +171,7 @@ pub(crate) fn is_named_accessor_in_table(class_or_builtin_ty: &TyName, godot_met
 
 /// Whether a class or builtin method should be hidden from the public API.
 #[rustfmt::skip]
-pub(crate) fn is_method_private(class_or_builtin_ty: &TyName, godot_method_name: &str) -> bool {
+pub fn is_method_private(class_or_builtin_ty: &TyName, godot_method_name: &str) -> bool {
     match (class_or_builtin_ty.godot_ty.as_str(), godot_method_name) {
         // Already covered by manual APIs
         | ("Object", "to_string")
@@ -183,7 +185,7 @@ pub(crate) fn is_method_private(class_or_builtin_ty: &TyName, godot_method_name:
 }
 
 #[rustfmt::skip]
-pub(crate) fn is_method_excluded_from_default_params(class_name: Option<&TyName>, godot_method_name: &str) -> bool {
+pub fn is_method_excluded_from_default_params(class_name: Option<&TyName>, godot_method_name: &str) -> bool {
     // None if global/utilities function
     let class_name = class_name.map_or("", |ty| ty.godot_ty.as_str());
 
@@ -200,7 +202,7 @@ pub(crate) fn is_method_excluded_from_default_params(class_name: Option<&TyName>
 /// should be returned to take precedence over general rules. Example: `FileAccess::get_pascal_string()` is mut, but would be const-qualified
 /// since it looks like a getter.
 #[rustfmt::skip]
-pub(crate) fn is_class_method_const(class_name: &TyName, godot_method: &JsonClassMethod) -> Option<bool> {
+pub fn is_class_method_const(class_name: &TyName, godot_method: &JsonClassMethod) -> Option<bool> {
     match (class_name.godot_ty.as_str(), godot_method.name.as_str()) {
         // Changed to const.
         | ("Object", "to_string")
@@ -239,33 +241,27 @@ pub(crate) fn is_class_method_const(class_name: &TyName, godot_method: &JsonClas
 }
 
 /// True if builtin method is excluded. Does NOT check for type exclusion; use [`is_builtin_type_deleted`] for that.
-pub(crate) fn is_builtin_method_deleted(_class_name: &TyName, method: &JsonBuiltinMethod) -> bool {
+pub fn is_builtin_method_deleted(_class_name: &TyName, method: &JsonBuiltinMethod) -> bool {
     // Currently only deleted if codegen.
     codegen_special_cases::is_builtin_method_excluded(method)
 }
 
 /// True if builtin type is excluded (`NIL` or scalars)
-pub(crate) fn is_builtin_type_deleted(class_name: &TyName) -> bool {
+pub fn is_builtin_type_deleted(class_name: &TyName) -> bool {
     let name = class_name.godot_ty.as_str();
     name == "Nil" || is_builtin_type_scalar(name)
 }
 
 /// True if `int`, `float`, `bool`, ...
-pub(crate) fn is_builtin_type_scalar(name: &str) -> bool {
+pub fn is_builtin_type_scalar(name: &str) -> bool {
     name.chars().next().unwrap().is_ascii_lowercase()
 }
 
-pub(crate) fn is_utility_function_deleted(
-    function: &JsonUtilityFunction,
-    ctx: &mut Context,
-) -> bool {
+pub fn is_utility_function_deleted(function: &JsonUtilityFunction, ctx: &mut Context) -> bool {
     codegen_special_cases::is_utility_function_excluded(function, ctx)
 }
 
-pub(crate) fn maybe_rename_class_method<'m>(
-    class_name: &TyName,
-    godot_method_name: &'m str,
-) -> &'m str {
+pub fn maybe_rename_class_method<'m>(class_name: &TyName, godot_method_name: &'m str) -> &'m str {
     match (class_name.godot_ty.as_str(), godot_method_name) {
         // GDScript, GDScriptNativeClass, possibly more in the future
         (_, "new") => "instantiate",
@@ -274,7 +270,7 @@ pub(crate) fn maybe_rename_class_method<'m>(
 }
 
 // Maybe merge with above?
-pub(crate) fn maybe_rename_virtual_method(rust_method_name: &str) -> &str {
+pub fn maybe_rename_virtual_method(rust_method_name: &str) -> &str {
     // A few classes define a virtual method called "_init" (distinct from the constructor)
     // -> rename those to avoid a name conflict in I* interface trait.
     match rust_method_name {
